+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
-- ...
+- Test interactors (used to synthetically manipulate state in tests)
+- CurrentPlayerCalculator helper
 
 ### Changed
 - ...